tests: Exclude expensive tests in check target by default.
This reduces the test time of 'make -j check' from ~22 minutes to a bit more than one
minute on a fast machine.

* Makefile.am (pre-check): New target.
(check): Add pre-check as a prerequisite.
* guix/tests.scm (with-slow-tests?): New variable
* tests/guix-home.sh
* tests/guix-package.sh
* tests/guix-system.sh
* tests/guix-build.sh
* tests/guix-gc.sh
* tests/guix-build-branch.sh
* tests/guix-refresh.sh: Skip unless WITH_SLOW_TESTS is 1.
* tests/guix-time-machine.sh: Exclude one command.
* tests/derivations.scm [!WITH_SLOW_TESTS]
("fixed-output derivation, network access, localhost): Skip test.

+++ tests/guix-gc.sh
@@ -20,6 +20,9 @@
 # Test the `guix gc' command-line utility.
 #
 
+# Slows the test suite by about half a minute.
+test "$WITH_SLOW_TESTS" != 1 && exit 77
+
 guix gc --version
 
 trap "rm -f guix-gc-root" EXIT
